refactor(react): update DatePicker to use cerberus config
