+++ packages/react/src/components/DatePicker.server.tsx
@@ -5,14 +5,10 @@ import {
   type DatePickerTableCellTriggerProps,
   type DatePickerTableHeaderProps,
   type DatePickerTableProps,
-  type DatePickerTriggerProps,
   type DatePickerViewControlProps,
 } from '@ark-ui/react'
-import { css, cx } from '@cerberus/styled-system/css'
-import { datePicker, iconButton, label } from '@cerberus/styled-system/recipes'
-import { $cerberusIcons } from '../config/defineIcons'
-import { IconButton } from './IconButton'
-import { Button } from './Button'
+import { cx } from '@cerberus/styled-system/css'
+import { datePicker, label } from '@cerberus/styled-system/recipes'
 
 /**
  * This module contains the DatePicker server family components.
@@ -48,33 +44,6 @@ export function DatePickerLabel(props: DatePickerLabelProps) {
 }
 
 /**
- * The trigger component for the DatePicker which opens the calendar.
- * @definition [datePicker docs](https://cerberus.digitalu.design/react/date-picker)
- * @example
- * ```tsx
- * <DatePickerTrigger />
- * ```
- */
-export function DatePickerTrigger(props: DatePickerTriggerProps) {
-  const { calendar: CalendarIcon } = $cerberusIcons
-  return (
-    <ArkDP.Trigger
-      {...props}
-      className={cx(
-        props.className,
-        iconButton({
-          size: 'sm',
-          usage: 'ghost',
-        }),
-        datePickerStyles.trigger,
-      )}
-    >
-      <CalendarIcon />
-    </ArkDP.Trigger>
-  )
-}
-
-/**
  * The control component for the DatePicker which wraps the triggers to switch
  * between calendar views.
  * @definition [datePicker docs](https://cerberus.digitalu.design/react/date-picker)
@@ -93,48 +62,6 @@ export function DatePickerViewControl(props: DatePickerViewControlProps) {
       {...props}
       className={cx(props.className, datePickerStyles.viewControl)}
     />
-  )
-}
-
-/**
- * An abstraction of the DatePicker control that contains the prev, view, and
- * next triggers which control the calendar output.
- * @definition [datePicker docs](https://cerberus.digitalu.design/react/date-picker)
- * @example
- * ```tsx
- * <DatePickerViewControlGroup />
- * ```
- */
-export function DatePickerViewControlGroup(props: DatePickerViewControlProps) {
-  const { calendarPrev: PrevIcon, calendarNext: NextIcon } = $cerberusIcons
-  return (
-    <DatePickerViewControl {...props}>
-      <ArkDP.PrevTrigger asChild>
-        <IconButton ariaLabel="Previous" size="sm">
-          <PrevIcon />
-        </IconButton>
-      </ArkDP.PrevTrigger>
-
-      <ArkDP.ViewTrigger asChild>
-        <Button
-          className={css({
-            h: '2rem',
-            paddingInline: 'md',
-          })}
-          shape="rounded"
-          size="sm"
-          usage="ghost"
-        >
-          <ArkDP.RangeText />
-        </Button>
-      </ArkDP.ViewTrigger>
-
-      <ArkDP.NextTrigger asChild>
-        <IconButton ariaLabel="Next" size="sm">
-          <NextIcon />
-        </IconButton>
-      </ArkDP.NextTrigger>
-    </DatePickerViewControl>
   )
 }
